feat(calc): add a temporary eval implementation

+++ CppCalc/CppCalc.h
@@ -9,5 +9,5 @@ public:
     CppCalc();
     ~CppCalc();
 
-    int64_t eval(const std::string &src);
+    int32_t eval(const std::string &src);
 };
